# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "joinedcontext Portal API",
    "description": "Administrative and platform management REST API for joinedcontext Portal",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/blueprints": {
      "get": {
        "tags": [
          "blueprints"
        ],
        "operationId": "list_blueprints",
        "responses": {
          "200": {
            "description": "The blueprints this caller may run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "operationId": "get_preferences",
        "responses": {
          "200": {
            "description": "The caller's preferences, empty before the first save",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No preferences database configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "operationId": "put_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stored; the body is what is now saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          },
          "400": {
            "description": "A field the Portal understands is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Missing or mismatched CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No preferences database configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/flows": {
      "post": {
        "tags": [
          "blueprints"
        ],
        "operationId": "start_flow",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project the flow creates resources in",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Change proposal opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Change"
                }
              }
            }
          },
          "400": {
            "description": "The parameters do not satisfy the blueprint's schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such blueprint for this caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The form was filled against another version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "The git forge is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync": {
      "get": {
        "tags": [
          "system"
        ],
        "operationId": "get_sync_status",
        "responses": {
          "200": {
            "description": "Status of the background Git mirror synchronization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Change": {
        "type": "object",
        "description": "The `Change` resource describing a proposed configuration update.",
        "required": [
          "apiVersion",
          "kind",
          "metadata",
          "status"
        ],
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ChangeMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ChangeStatus"
          }
        }
      },
...
      "ChangeMeta": {
        "type": "object",
        "description": "Metadata identifying the change.",
        "required": [
          "name",
          "namespace"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Identifier formatted as `chg-` plus eight lowercase hex characters derived\nfrom the merge request."
          },
          "namespace": {
            "type": "string"
          }
        }
      },
      "ChangePhase": {
        "type": "string",
        "description": "Lifecycle phase of a change.",
        "enum": [
          "PendingApproval",
          "Deploying",
          "Merged",
          "Applied",
          "Rejected"
        ]
      },
...
      "ChangeStatus": {
        "type": "object",
        "description": "Approval and reconciliation status of a change.",
        "required": [
          "lane",
          "plan",
          "phase"
        ],
        "properties": {
          "lane": {
            "$ref": "#/components/schemas/Lane"
          },
          "mergeRequest": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/ChangePhase"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSummary"
          }
        }
      },
...
      "Condition": {
        "type": "object",
        "title": "Condition",
        "description": "Status condition entry.",
        "required": [
          "status",
          "type"
        ],
        "properties": {
          "lastTransitionTime": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Last transition timestamp."
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable message."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Machine-readable reason code."
          },
          "status": {
            "type": "string",
            "description": "Status value (\"True\", \"False\", \"Unknown\")."
          },
          "type": {
            "type": "string",
            "description": "Type of condition (e.g. Reconciled, Ready)."
          }
        },
        "additionalProperties": false
      },
...
      "FlowRequest": {
        "type": "object",
        "description": "Running a blueprint: the parameters the form collected (API/01 §13).",
        "required": [
          "blueprint",
          "version",
          "parameters"
        ],
        "properties": {
          "blueprint": {
            "type": "string",
            "description": "The organization-level Blueprint to run."
          },
          "parameters": {
            "description": "The values the user filled in, validated against `spec.parameterSchema` (CC-24)."
          },
          "version": {
            "type": "string",
            "description": "The version the form was generated from. A mismatch is a conflict rather than an\nexpansion against a schema the user never saw (CC-26)."
          }
        },
        "additionalProperties": false
      },
...
      "Lane": {
        "type": "string",
        "description": "Change risk classification lane (CC-63).",
        "enum": [
          "green",
          "yellow",
          "red"
        ]
      },
...
      "ListMeta": {
        "type": "object",
        "properties": {
          "continue": {
            "type": [
              "string",
              "null"
            ]
          },
          "remainingItemCount": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
...
      "ObjectMeta": {
        "type": "object",
        "title": "ObjectMeta",
        "description": "Metadata envelope attached to every platform resource (MF-02).",
        "required": [
          "name"
        ],
        "properties": {
          "annotations": {
            "type": "object",
            "description": "Key-value annotations for provenance and ownership (MF-08).",
            "additionalProperties": {
              "type": "string"
            }
          },
          "description": {
            "type": [
              "object",
              "null"
            ],
            "description": "Human-readable multilingual description (PF-24).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string",
              "pattern": "^[a-z]{2}$"
            }
          },
          "labels": {
            "type": "object",
            "description": "Key-value labels for filtering (MF-10).",
            "additionalProperties": {
              "type": "string"
            }
          },
          "name": {
            "type": "string",
            "description": "Resource name (DNS-1123 label)."
          },
          "namespace": {
            "type": [
              "string",
              "null"
            ],
            "description": "Resource namespace (\"org\" or project slug)."
          },
          "title": {
            "type": [
              "object",
              "null"
            ],
            "description": "Human-readable multilingual title (PF-24).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string",
              "pattern": "^[a-z]{2}$"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Phase": {
        "oneOf": [
          {
            "type": "string",
            "description": "Resource manifest is in draft state.",
            "enum": [
              "Draft"
            ]
          },
          {
            "type": "string",
            "description": "Resource is pending review or approval.",
            "enum": [
              "Pending"
            ]
          },
          {
            "type": "string",
            "description": "Resource deployment is in progress.",
            "enum": [
              "Deploying"
            ]
          },
          {
            "type": "string",
            "description": "Resource is live and active in cluster.",
            "enum": [
              "Live"
            ]
          },
          {
            "type": "string",
            "description": "Reconciliation error.",
            "enum": [
              "Error"
            ]
          }
        ],
        "title": "Phase",
        "description": "Lifecycle phase enumeration."
      },
...
      "PlanSummary": {
        "type": "object",
        "description": "Counts of planned resource mutations.",
        "properties": {
          "create": {
            "type": "integer",
            "minimum": 0
          },
          "delete": {
            "type": "integer",
            "minimum": 0
          },
          "update": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Preferences": {
        "type": "object",
        "properties": {
          "dashboardLayouts": {
            "type": "object",
            "description": "Free JSON per dashboard name, the browser's own saved state."
          },
          "defaultProject": {
            "type": [
              "string",
              "null"
            ],
            "description": "The project the shell opens on."
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 639-1 language code."
          },
          "theme": {
            "type": [
              "string",
              "null"
            ],
            "description": "`light`, `dark` or `system`."
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details representation.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every violation of one request, when there is more than one thing to say. A form marks\nall its bad fields from this in one pass instead of one round trip per mistake (CC-24)."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ResourceEnvelope": {
        "type": "object",
        "required": [
          "apiVersion",
          "kind",
          "metadata"
        ],
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ObjectMeta"
          },
          "spec": {
            "type": "object"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Status",
                "description": "Never read from Git, always computed by the Portal API (MF-04)."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ResourceList": {
        "type": "object",
        "required": [
          "apiVersion",
          "kind",
          "metadata",
          "items"
        ],
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceEnvelope"
            }
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ListMeta"
          }
        }
      },
...
      "Status": {
        "type": "object",
        "description": "The status the Portal API reports (MF-04). It is `jc_core::Status` plus `sourceUrl` and a phase\nthat is always known; it stays a Portal type until jc-core carries `sourceUrl` too (docs API/01\nsection 6 is the contract), then it becomes a re-export like its neighbours.",
        "required": [
          "phase"
        ],
        "properties": {
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Condition"
            }
          },
          "observedRevision": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "sourceUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Forge page of the file this manifest was read from, so a view can link \"Source\"\nwithout knowing where a kind lives in the repository. Computed, never read from Git."
          }
        },
        "additionalProperties": false
      },
...
      "SyncStatus": {
        "type": "object",
        "description": "Status of the background Git mirror synchronization.\n\nServed to the browser, so it deliberately never carries a token,\nrepository URL, or branch name.",
        "required": [
          "manifests",
          "leader"
        ],
        "properties": {
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastSync": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "leader": {
            "type": "boolean",
            "description": "Whether this replica is the one that reconciles (CC-03). A Portal without a database\nhas no election to run and reconciles on its own, so it reports itself as the leader."
          },
          "manifests": {
            "type": "integer",
            "minimum": 0
          },
          "revision": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  },
...
}